moved write_wing_FFD_file to a different file
